requirements.txt
requests
tqdm
python-dotenv
sentence-transformers
torch
scikit-learn
numpy
pandas
